requirements.txt
watchdog
pdfminer.six
requests
python-dotenv
google-generativeai
python-docx
striprtf 